stb_vorbis.h: fix gcc sign-compare warnings when building with -Wextra
Fixes https://github.com/libsdl-org/SDL_mixer/issues/614

+++ src/codecs/stb_vorbis/stb_vorbis.h
@@ -3766,7 +3766,7 @@ static int start_decoder(vorb *f)
    f->comment_list = NULL;
    if (f->comment_list_length > 0)
    {
-      if (INT_MAX / sizeof(char*) < f->comment_list_length)
+      if (INT_MAX / (int)sizeof(char*) < f->comment_list_length)
           goto no_comment;
       len = sizeof(char*) * f->comment_list_length;
       f->comment_list = (char**) setup_malloc(f, len);
